requirements.txt
azure-functions
openai
azure-cosmos
azure-identity
python-dotenv